chore: use upgrade by default and add --legacy flag
