+++ packages/cli/src/upgrade/upgrade.js
@@ -1,168 +1,210 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @format
- * @flow
- */
+// @flow
 /* eslint-disable consistent-return */
-
-import fs from 'fs';
 import path from 'path';
+import fs from 'fs';
 import semver from 'semver';
+import execa from 'execa';
 import type { ContextT } from '../core/types.flow';
 import logger from '../util/logger';
-import copyProjectTemplateAndReplace from '../generator/copyProjectTemplateAndReplace';
-import newUpgrade from './newUpgrade';
+import PackageManager from '../util/PackageManager';
+import { fetch } from './helpers';
+import legacyUpgrade from './legacyUpgrade';
 
 type FlagsT = {
-  improved: boolean,
+  legacy: boolean,
+};
+
+const getLatestRNVersion = async (): Promise<string> => {
+  logger.info('No version passed. Fetching latest...');
+  const { stdout } = await execa('npm', ['info', 'react-native', 'version']);
+  return stdout;
+};
+
+const getRNPeerDeps = async (
+  version: string
+): Promise<{ [key: string]: string }> => {
+  const { stdout } = await execa('npm', [
+    'info',
+    `react-native@${version}`,
+    'peerDependencies',
+    '--json',
+  ]);
+
+  return JSON.parse(stdout);
+};
+
+const getPatch = async (currentVersion, newVersion, projectDir) => {
+  let patch;
+  const rnDiffPurgeUrl = 'https://github.com/pvinis/rn-diff-purge';
+  const rnDiffAppName = 'RnDiffApp';
+  const { name } = require(path.join(projectDir, 'package.json'));
+
+  logger.info(`Fetching diff between v${currentVersion} and v${newVersion}...`);
+
+  try {
+    patch = await fetch(
+      `${rnDiffPurgeUrl}/compare/version/${currentVersion}...version/${newVersion}.diff`
+    );
+  } catch (error) {
+    logger.error(
+      `Failed to fetch diff for react-native@${newVersion}. Maybe it's not released yet?`
+    );
+    logger.info(
+      `For available releases to diff see: https://github.com/pvinis/rn-diff-purge#version-changes`
+    );
+    return null;
+  }
+
+  return patch
+    .replace(new RegExp(rnDiffAppName, 'g'), name)
+    .replace(new RegExp(rnDiffAppName.toLowerCase(), 'g'), name.toLowerCase());
+};
+
+const getVersionToUpgradeTo = async (argv, currentVersion, projectDir) => {
+  const newVersion = argv[0]
+    ? semver.valid(argv[0]) ||
+      (semver.coerce(argv[0]) ? semver.coerce(argv[0]).version : null)
+    : await getLatestRNVersion();
+
+  if (!newVersion) {
+    logger.error(
+      `Provided version "${
+        argv[0]
+      }" is not allowed. Please pass a valid semver version`
+    );
+    return null;
+  }
+
+  if (currentVersion > newVersion) {
+    logger.error(
+      `Trying to upgrade from newer version "${currentVersion}" to older "${newVersion}"`
+    );
+    return null;
+  }
+  if (currentVersion === newVersion) {
+    const {
+      dependencies: { 'react-native': version },
+    } = require(path.join(projectDir, 'package.json'));
+
+    if (semver.satisfies(newVersion, version)) {
+      logger.warn(
+        `Specified version "${newVersion}" is already installed in node_modules and it satisfies "${version}" semver range. No need to upgrade`
+      );
+      return null;
+    }
+    logger.error(
+      `Dependency mismatch. Specified version "${newVersion}" is already installed in node_modules and it doesn't satisfy "${version}" semver range of your "react-native" dependency. Please re-install your dependencies`
+    );
+    return null;
+  }
+
+  return newVersion;
+};
+
+const installDeps = async (newVersion, projectDir) => {
+  logger.info(
+    `Installing react-native@${newVersion} and its peer dependencies...`
+  );
+  const peerDeps = await getRNPeerDeps(newVersion);
+  const pm = new PackageManager({ projectDir });
+  const deps = [
+    `react-native@${newVersion}`,
+    ...Object.keys(peerDeps).map(module => `${module}@${peerDeps[module]}`),
+  ];
+  pm.install(deps);
 };
 
 /**
- * Migrate application to a new version of React Native.
- * See http://facebook.github.io/react-native/docs/upgrading.html
+ * Upgrade application to a new version of React Native.
  */
-function validateAndUpgrade(argv: Array<string>, ctx: ContextT, args: FlagsT) {
-  if (args.improved) {
-    return newUpgrade.func(argv, ctx);
+async function upgrade(argv: Array<string>, ctx: ContextT, args: FlagsT) {
+  if (args.legacy) {
+    return legacyUpgrade.func(argv, ctx);
   }
+  const rnDiffGitAddress = `https://github.com/pvinis/rn-diff-purge.git`;
+  const tmpRemote = 'tmp-rn-diff-purge';
+  const tmpPatchFile = 'tmp-upgrade-rn.patch';
   const projectDir = ctx.root;
-
-  const packageJSON = JSON.parse(
-    fs.readFileSync(path.resolve(projectDir, 'package.json'), 'utf8')
-  );
-
-  warn(
-    'You should consider using the new upgrade tool based on Git. It ' +
-      'makes upgrades easier by resolving most conflicts automatically.\n' +
-      'To use it:\n' +
-      '- Go back to the old version of React Native\n' +
-      '- Run "npm install -g react-native-git-upgrade"\n' +
-      '- Run "react-native-git-upgrade"\n' +
-      'See https://facebook.github.io/react-native/docs/upgrading.html'
-  );
-
-  const projectName = packageJSON.name;
-  if (!projectName) {
-    warn(
-      'Your project needs to have a name, declared in package.json, ' +
-        'such as "name": "AwesomeApp". Please add a project name. Aborting.'
-    );
-    return;
-  }
-
-  const version = packageJSON.dependencies['react-native'];
-  if (!version) {
-    warn(
-      'Your "package.json" file doesn\'t seem to declare "react-native" as ' +
-        'a dependency. Nothing to upgrade. Aborting.'
-    );
-    return;
-  }
-
-  if (version === 'latest' || version === '*') {
-    warn(
-      'Some major releases introduce breaking changes.\n' +
-        'Please use a caret version number in your "package.json" file \n' +
-        'to avoid breakage. Use e.g. react-native: ^0.38.0. Aborting.'
-    );
-    return;
-  }
-
-  const installed = JSON.parse(
-    fs.readFileSync(
-      path.resolve(projectDir, 'node_modules/react-native/package.json'),
-      'utf8'
-    )
-  );
-
-  if (!semver.satisfies(installed.version, version)) {
-    warn(
-      'react-native version in "package.json" doesn\'t match ' +
-        'the installed version in "node_modules".\n' +
-        'Try running "npm install" to fix this. Aborting.'
-    );
-    return;
-  }
-
-  const v = version.replace(/^(~|\^|=)/, '').replace(/x/i, '0');
-
-  if (!semver.valid(v)) {
-    warn(
-      "A valid version number for 'react-native' is not specified in your " +
-        "'package.json' file. Aborting."
-    );
-    return;
-  }
-
-  logger.info(
-    `Upgrading project to react-native v${installed.version}\n` +
-      `Check out the release notes and breaking changes: ` +
-      `https://github.com/facebook/react-native/releases/tag/v${semver.major(
-        v
-      )}.${semver.minor(v)}.0`
-  );
-
-  // >= v0.21.0, we require react to be a peer dependency
-  if (semver.gte(v, '0.21.0') && !packageJSON.dependencies.react) {
-    warn(
-      'Your "package.json" file doesn\'t seem to have "react" as a dependency.\n' +
-        '"react" was changed from a dependency to a peer dependency in react-native v0.21.0.\n' +
-        'Therefore, it\'s necessary to include "react" in your project\'s dependencies.\n' +
-        'Please run "npm install --save react", then re-run "react-native upgrade".\n'
-    );
-    return;
-  }
-
-  if (semver.satisfies(v, '~0.26.0')) {
-    warn(
-      'React Native 0.26 introduced some breaking changes to the native files on iOS. You can\n' +
-        'perform them manually by checking the release notes or use "rnpm" ' +
-        'to do it automatically.\n' +
-        'Just run:\n' +
-        '"npm install -g rnpm && npm install rnpm-plugin-upgrade@0.26 --save-dev", ' +
-        'then run "rnpm upgrade".'
-    );
-  }
-
-  upgradeProjectFiles(projectDir, projectName);
-
-  logger.info(
-    `Successfully upgraded this project to react-native v${installed.version}`
-  );
-}
-
-/**
- * Once all checks passed, upgrade the project files.
- */
-function upgradeProjectFiles(projectDir, projectName) {
-  // Just overwrite
-  copyProjectTemplateAndReplace(
-    path.dirname(require.resolve('react-native/template')),
+  const { version: currentVersion } = require(path.join(
     projectDir,
-    projectName,
-    { upgrade: true }
-  );
-}
+    'node_modules/react-native/package.json'
+  ));
 
-function warn(message) {
-  logger.warn(message);
+  const newVersion = await getVersionToUpgradeTo(
+    argv,
+    currentVersion,
+    projectDir
+  );
+
+  if (!newVersion) {
+    return;
+  }
+
+  const patch = await getPatch(currentVersion, newVersion, projectDir);
+
+  if (patch === null) {
+    return;
+  }
+
+  if (patch === '') {
+    logger.info('Diff has no changes to apply, proceeding further');
+    await installDeps(newVersion, projectDir);
+    logger.success(
+      `Upgraded React Native to v${newVersion} 🎉. Now you can review and commit the changes`
+    );
+    return;
+  }
+
+  try {
+    fs.writeFileSync(tmpPatchFile, patch);
+    await execa('git', ['remote', 'add', tmpRemote, rnDiffGitAddress]);
+    await execa('git', ['fetch', tmpRemote]);
+
+    try {
+      logger.info('Applying diff...');
+      await execa(
+        'git',
+        ['apply', tmpPatchFile, '--exclude=package.json', '-p2', '--3way'],
+        { stdio: 'inherit' }
+      );
+    } catch (error) {
+      logger.error(
+        `Applying diff failed. Please review the conflicts and resolve them.`
+      );
+      logger.info(
+        `You may find release notes helpful: https://github.com/facebook/react-native/releases/tag/v${newVersion}`
+      );
+      return;
+    }
+
+    await installDeps(newVersion, projectDir);
+  } catch (error) {
+    throw new Error(error.stderr || error);
+  } finally {
+    try {
+      fs.unlinkSync(tmpPatchFile);
+    } catch (e) {
+      // ignore
+    }
+    await execa('git', ['remote', 'remove', tmpRemote]);
+  }
+
+  logger.success(
+    `Upgraded React Native to v${newVersion} 🎉. Now you can review and commit the changes`
+  );
 }
 
 const upgradeCommand = {
-  name: 'upgrade',
+  name: 'upgrade [version]',
   description:
-    "upgrade your app's template files to the latest version; run this after " +
-    'updating the react-native version in your package.json and running npm install',
-  func: validateAndUpgrade,
+    "Upgrade your app's template files to the specified or latest npm version using `rn-diff-purge` project. Only valid semver versions are allowed.",
+  func: upgrade,
   options: [
     {
-      command: '--improved',
+      command: '--legacy',
       description:
-        'Improved implementation using rn-diff-purge project for diffing. This is going to be the default in next major and the old way of upgrading will be deprecated',
+        "Legacy implementation. Upgrade your app's template files to the latest version; run this after " +
+        'updating the react-native version in your package.json and running npm install',
     },
   ],
 };
